+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - **Fix Table rows rendering when selection is enabled and min width is set for columns**
 
+### What's new
+
+- **Added `UserIconGroup` component** - stacked/non-stacked, animated lists with custom count icon behavior.
+
 ## [1.11.0]
 
 `2021-07-08`